Introduce module loader abstraction
