+++ tests/lazy_stdlib.rs
@@ -53,3 +53,10 @@ fn nested_stats_imports_work_and_require_imports() {
     assert!(run("+> stats\nx := stats/sum/column([1 2 3])"));
     assert!(run("+> stats/sum/column\nx := column([1 2 3])"));
 }
+
+#[cfg(feature = "linked_stdlib")]
+#[test]
+fn repeated_module_and_item_imports_remain_idempotent() {
+    assert!(run("+> math\n+> math\nx := math/sin(1.23)"));
+    assert!(run("+> math\n+> math/sin\nx := sin(1.23)"));
+}
